+++ src/lib/path.ts
@@ -10,7 +10,7 @@ type PathValueRecursive<T, P extends string> = P extends ''
   ? T
   : P extends `.${infer Key}[${number}]${infer Rest}`
     ? Key extends keyof T
-      ? T[Key] extends (infer U)[]
+      ? T[Key] extends readonly (infer U)[]
         ? PathValueRecursive<U, Rest>
         : unknown
       : P extends `.${infer Key2}.${infer Rest2}`
@@ -19,7 +19,7 @@ type PathValueRecursive<T, P extends string> = P extends ''
           : unknown
         : unknown
     : P extends `[${number}]${infer Rest}`
-      ? T extends (infer U)[]
+      ? T extends readonly (infer U)[]
         ? PathValueRecursive<U, Rest>
         : unknown
       : P extends `.${infer Key}.${infer Rest}`